svncmds: Correct colored diff output
[Why]
On linux if colordiff is installed the diff output is broken, as the
svn diff output already contains the color coding. Therefor the color
output does not work as well, since the line does not start with the
expected symbols.

[How]
Force svn diff to use the internal diff instead a possible external
tool.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>